Split entity test cases to multiple files correct the signature of method `Adapter._filter_objects`
