+++ tests/core/test_entity.py
@@ -27,7 +27,6 @@ from protean.core.entity import Entity
 from protean.core.entity import QuerySet
 from protean.core.exceptions import ObjectNotFoundError
 from protean.core.exceptions import ValidationError
-from protean.utils.query import Q
 
 
 class TestEntity:
@@ -515,84 +514,9 @@ class TestEntity:
         query = Dog.query.filter(owner='John')
         assert isinstance(query, QuerySet)
 
-    class TestState:
-        """Class that holds tests for Entity State Management"""
-
-        def test_default_state(self):
-            """Test that a default state is available when the entity is instantiated"""
-            dog = Dog(id=1, name='John Doe', age=10, owner='Jimmy')
-            assert dog._state is not None
-            assert dog._state._new
-            assert dog._state.is_new()
-            assert dog.is_new
-            assert not dog.is_persisted
-
-        def test_state_on_retrieved_objects(self):
-            """Test that retrieved objects are not marked as new"""
-            dog = Dog.create(name='John Doe', age=10, owner='Jimmy')
-            dog_dup = Dog.get(dog.id)
-
-            assert not dog_dup.is_new
-
-        def test_persisted_after_save(self):
-            """Test that the entity is marked as saved after successfull save"""
-            dog = Dog(id=1, name='John Doe', age=10, owner='Jimmy')
-            assert dog.is_new
-            dog.save()
-            assert dog.is_persisted
-
-        def test_not_persisted_if_save_failed(self):
-            """Test that the entity still shows as new if save failed"""
-            dog = Dog(id=1, name='John Doe', age=10, owner='Jimmy')
-            try:
-                del dog.name
-                dog.save()
-            except ValidationError as exc:
-                assert dog.is_new
-
-        def test_persisted_after_create(self):
-            """Test that the entity is marked as saved after successfull create"""
-            dog = Dog.create(id=1, name='John Doe', age=10, owner='Jimmy')
-            assert not dog.is_new
-
-        def test_copy_resets_state(self):
-            """Test that a default state is available when the entity is instantiated"""
-            dog1 = Dog.create(id=1, name='John Doe', age=10, owner='Jimmy')
-            dog2 = dog1.clone()
-
-            assert dog2.is_new
-
-        def test_changed(self):
-            """Test that entity is marked as changed if attributes are updated"""
-            dog = Dog.create(id=1, name='John Doe', age=10, owner='Jimmy')
-            assert not dog.is_changed
-            dog.name = 'Jane Doe'
-            assert dog.is_changed
-
-        def test_not_changed_if_still_new(self):
-            """Test that entity is not marked as changed upon attribute change if its still new"""
-            dog = Dog(id=1, name='John Doe', age=10, owner='Jimmy')
-            assert not dog.is_changed
-            dog.name = 'Jane Doe'
-            assert not dog.is_changed
-
-        def test_not_changed_after_save(self):
-            """Test that entity is marked as not changed after save"""
-            dog = Dog.create(id=1, name='John Doe', age=10, owner='Jimmy')
-            dog.name = 'Jane Doe'
-            assert dog.is_changed
-            dog.save()
-            assert not dog.is_changed
-
-        def test_destroyed(self):
-            """Test that a entity is marked as destroyed after delete"""
-            dog = Dog.create(id=1, name='John Doe', age=10, owner='Jimmy')
-            assert not dog.is_destroyed
-            dog.delete()
-            assert dog.is_destroyed
-
 
 class TestEntityMetaAttributes:
+
     """Class that holds testcases for Entity's meta attributes"""
 
     def test_meta_on_init(self):
@@ -624,996 +548,3 @@ class TestEntityMetaAttributes:
 
         assert all(key in dog.attributes for key in ['age', 'has_one_human1_id', 'id', 'name'])
         assert all(key in human.attributes for key in ['first_name', 'id', 'last_name', 'email'])
-
-
-class TestQuerySet:
-    """Class that holds Tests for QuerySet"""
-
-    def test_clone(self):
-        """Test that clone works as expected... it clones!"""
-        query1 = Dog.query.filter(owner='John')
-        query2 = query1.filter(age=3)
-        query3 = query2.order_by('name')
-
-        assert query1 != query2
-        assert query2 != query3
-
-    def test_list(self):
-        """Test that filter is evaluted on calling `list()`"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-        dogs = list(query)
-
-        assert dogs is not None
-        assert len(dogs) == 2
-
-    def test_repr(self):
-        """Test that filter is evaluted on calling `list()`"""
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert repr(query) == ("<QuerySet: entity: Dog, "
-                               "criteria: ('protean.utils.query.Q', (), {'owner': 'John'}), "
-                               "page: 1, "
-                               "per_page: 10, order_by: {'age'}>")
-
-    def test_bool_false(self):
-        """Test that `bool` returns `False` on no records"""
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert bool(query) is False
-
-    def test_bool_true(self):
-        """Test that filter is evaluted on calling `list()`"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-
-        assert bool(query) is True
-
-    def test_len(self):
-        """Test that filter is evaluted on calling `list()`"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert len(query) == 2
-
-    def test_slice(self):
-        """Test slicing on filter"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-        Dog.create(id=5, name='Fred', age=4, owner='Constantine')
-        Dog.create(id=6, name='Flint', age=2, owner='Steve')
-
-        # Filter by Dog attributes
-        query = Dog.query.order_by('age')
-        sliced = query[1:]
-        assert len(sliced) == 4
-
-    def test_caching(self):
-        """Test that results are cached after query is evaluated"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-
-        # Result cache is empty to begin with
-        assert query._result_cache is None
-
-        # Total invokes an evaluation and a query
-        assert query.total == 2
-
-        # Result cache is now populated
-        assert query._result_cache is not None
-        assert query._result_cache.total == 2
-
-    def test_cache_reset(self):
-        """Test that results are cached after query is evaluated"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-
-        # Total invokes an evaluation and a query
-        assert query.total == 2
-        assert query._result_cache.total == 2
-
-        query_dup = query.paginate(per_page=25)
-        assert query_dup._result_cache is None
-
-    def test_total(self):
-        """Test value of `total` results"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert query.total == 2
-
-    def test_total_with_cache(self):
-        """Test value of `total` results without refresh"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert query.total == 2
-
-        Dog.create(id=5, name='Berry', age=1, owner='John')
-        assert query.total == 2
-        assert query._result_cache.total == 2
-
-        # Force a refresh
-        assert query.all().total == 3
-
-        # Result cache is now populated
-        assert query._result_cache.total == 3
-
-    def test_items(self):
-        """Test that items is retrieved from Pagination results"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert query.items[0].id == query.all().items[0].id
-
-    def test_items_with_cache(self):
-        """Test that items is retrieved from Pagination results"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.filter(owner='John').order_by('age')
-        assert query.items[0].id == 3
-
-        Dog.create(id=5, name='Berry', age=1, owner='John')
-        assert query.items[0].id == 3
-
-        assert query.all().items[0].id == 5
-
-    def test_has_next(self):
-        """Test if there are results after the current set"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.paginate(page=1, per_page=2)
-        assert query.has_next is True
-
-    def test_has_next_with_cache(self):
-        """Test if there are results after the current set"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        dog = Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.paginate(page=1, per_page=2)
-        assert query.has_next is True
-
-        dog.delete()
-
-        assert query.has_next is True
-        assert query.all().has_next is False
-
-    def test_has_prev(self):
-        """Test if there are results before the current set"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.paginate(page=2, per_page=2)
-        assert query.has_prev is True
-
-    def test_has_prev_with_cache(self):
-        """Test if there are results before the current set"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        dog = Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.paginate(page=2, per_page=2)
-        assert query.has_prev is True
-
-        dog.delete()
-
-        assert query.has_prev is True
-        assert query.all().has_prev is False
-
-    def test_first(self):
-        """Test that the first item is retrieved correctly from the resultset"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.order_by('-age')
-        assert query.first.id == 2
-
-    def test_first_with_cache(self):
-        """Test that the first item is retrieved correctly from the resultset"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by Dog attributes
-        query = Dog.query.order_by('-age')
-        assert query.first.id == 2
-
-        Dog.create(id=5, name='Berry', age=8, owner='John')
-        assert query.first.id == 2
-        assert query.all().first.id == 5
-
-
-class TestCriteriaConstruction:
-    """Test Conjunction operations on QuerySet"""
-
-    def test_empty_query(self):
-        """Test that an empty Q object is initialized with Queryset"""
-        assert isinstance(Dog.query, QuerySet)
-        assert isinstance(Dog.query._criteria, Q)
-
-    def test_simple_filter(self):
-        """Test query construction with simple filter kwargs"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(owner='John')
-        assert q1._criteria is not None
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert decon_args == ()
-        assert decon_kwargs == {'owner': 'John'}
-        assert q1._criteria.connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_simple_exclude(self):
-        """Test query construction with simple exclude kwargs"""
-        # Filter by the Owner
-        q1 = Dog.query.exclude(owner='John')
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert decon_args == ()
-        assert decon_kwargs == {'owner': 'John', '_negated': True}
-        assert q1._criteria.connector == Q.AND
-        assert q1._criteria.negated is True
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_multiple_filters(self):
-        """Test query construction with multiple filter criteria"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(owner='John')
-        q2 = q1.filter(age=3)
-
-        _, decon_args, decon_kwargs = q2._criteria.deconstruct()
-
-        assert decon_args == (('owner', 'John'), ('age', 3))
-        assert decon_kwargs == {}
-        assert q1._criteria.connector == Q.AND
-        assert q2._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q2._criteria
-
-    def test_multiple_excludes(self):
-        """Test query construction with multiple exclude criteria"""
-        # Filter by the Owner
-        q1 = Dog.query.exclude(owner='John')
-        q2 = q1.exclude(age=3)
-
-        _, decon_args, decon_kwargs = q2._criteria.deconstruct()
-
-        assert str(decon_args) == ("(<Q: (NOT (AND: ('owner', 'John')))>, "
-                                   "<Q: (NOT (AND: ('age', 3)))>)")
-        assert decon_kwargs == {}
-        assert q1._criteria.connector == Q.AND
-        assert q2._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q2._criteria
-
-    def test_multiple_criteria_in_filter(self):
-        """Test query construction with multiple filter in filter"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(owner='John', age=3)
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert decon_args == (('age', 3), ('owner', 'John'))
-        assert decon_kwargs == {}
-        assert q1._criteria.connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_multiple_criteria_in_exclude(self):
-        """Test query construction with multiple filter in exclude"""
-        # Filter by the Owner
-        q1 = Dog.query.exclude(owner='John', age=3)
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert decon_args == (('age', 3), ('owner', 'John'))
-        assert decon_kwargs == {'_negated': True}
-        assert q1._criteria.connector == Q.AND
-        assert q1._criteria.negated is True
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_combined_filter_and_exclude(self):
-        """Test query construction with combined filter/exclude with filter coming first"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(owner='John').exclude(age=3)
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(('owner', 'John'), <Q: (NOT (AND: ('age', 3)))>)"
-        assert decon_kwargs == {}
-        assert q1._criteria.connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_combined_exclude_and_filter(self):
-        """Test query construction with combined filter/exclude with exclude coming first"""
-        # Filter by the Owner
-        q1 = Dog.query.exclude(age=3).filter(owner='John')
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (NOT (AND: ('age', 3)))>, ('owner', 'John'))"
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_single_Q_object(self):
-        """Test query construction with single Q instance as input to `filter` method"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John'))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (AND: ('owner', 'John'))>,)"
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_exclude_with_single_Q_object(self):
-        """Test query construction with single Q instance as input to `exclude` method"""
-        # Filter by the Owner
-        q1 = Dog.query.exclude(Q(owner='John'))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (AND: ('owner', 'John'))>,)"
-        assert decon_kwargs == {'_negated': True}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is True
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_multiple_Q_objects(self):
-        """Test query construction with multiple Q objects input to `filter` method"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John'), Q(age=3))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (AND: ('owner', 'John'))>, <Q: (AND: ('age', 3))>)"
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_exclude_with_multiple_Q_objects(self):
-        """Test query construction with multiple Q objects input to `exclude` method"""
-        q1 = Dog.query.exclude(Q(owner='John'), Q(age=3))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (AND: ('owner', 'John'))>, <Q: (AND: ('age', 3))>)"
-        assert decon_kwargs == {'_negated': True}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is True
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_AND(self):
-        """Test query construction with AND"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John') & Q(age=3))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (AND: ('owner', 'John'), ('age', 3))>,)"
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_OR(self):
-        """Test query construction with OR"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John') | Q(age=3))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == "(<Q: (OR: ('owner', 'John'), ('age', 3))>,)"
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.OR
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_multiple_ANDs(self):
-        """Test query construction with multiple AND criteria"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John') & Q(age=3) & Q(name='Jean'))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == ("(<Q: (AND: "
-                                   "('owner', 'John'), "
-                                   "('age', 3), "
-                                   "('name', 'Jean'))>,)")
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.AND
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_multiple_ORs(self):
-        """Test query construction with multiple OR criteria"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John') | Q(age=3) | Q(name='Jean'))
-
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-
-        assert str(decon_args) == ("(<Q: (OR: "
-                                   "('owner', 'John'), "
-                                   "('age', 3), "
-                                   "('name', 'Jean'))>,)")
-        assert decon_kwargs == {}
-        assert q1._criteria.children[0].connector == Q.OR
-        assert q1._criteria.negated is False
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-
-    def test_filter_with_AND_OR_combination(self):
-        """Test query construction with AND and OR combinations"""
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John') | Q(age=3), name='Jean')
-        _, decon_args, decon_kwargs = q1._criteria.deconstruct()
-        assert str(decon_args) == ("(<Q: (OR: ('owner', 'John'), ('age', 3))>, "
-                                   "('name', 'Jean'))")
-        assert Q(*decon_args, **decon_kwargs) == q1._criteria
-        assert q1._criteria.children[0].connector == Q.OR
-
-        q2 = Dog.query.filter((Q(owner='John') | Q(age=3)), Q(name='Jean'))
-        _, decon_args, decon_kwargs = q2._criteria.deconstruct()
-        assert str(decon_args) == ("(<Q: (OR: ('owner', 'John'), ('age', 3))>, "
-                                   "<Q: (AND: ('name', 'Jean'))>)")
-        assert Q(*decon_args, **decon_kwargs) == q2._criteria
-        assert q2._criteria.children[0].connector == Q.OR
-
-        q3 = Dog.query.filter(Q(name='Jean') & (Q(owner='John') | Q(age=3)))
-        _, decon_args, decon_kwargs = q3._criteria.deconstruct()
-        assert str(decon_args) == ("(<Q: (AND: "
-                                   "('name', 'Jean'), "
-                                   "(OR: ('owner', 'John'), ('age', 3)))>,)")
-        assert Q(*decon_args, **decon_kwargs) == q3._criteria
-        assert q3._criteria.children[0].connector == Q.AND
-
-
-class TestQ:
-    """Class that holds tests for Q Objects"""
-
-    def test_deconstruct(self):
-        q = Q(price__gt=10.0)
-        path, args, kwargs = q.deconstruct()
-        assert path == 'protean.utils.query.Q'
-        assert args == ()
-        assert kwargs == {'price__gt': 10.0}
-
-    def test_deconstruct_negated(self):
-        q = ~Q(price__gt=10.0)
-        path, args, kwargs = q.deconstruct()
-        assert args == ()
-        assert kwargs == {
-            'price__gt': 10.0,
-            '_negated': True,
-        }
-
-    def test_deconstruct_or(self):
-        q1 = Q(price__gt=10.0)
-        q2 = Q(price=11.0)
-        q3 = q1 | q2
-        path, args, kwargs = q3.deconstruct()
-        assert args == (
-            ('price__gt', 10.0),
-            ('price', 11.0),
-        )
-        assert kwargs == {'_connector': 'OR'}
-
-    def test_deconstruct_and(self):
-        q1 = Q(price__gt=10.0)
-        q2 = Q(price=11.0)
-        q = q1 & q2
-        path, args, kwargs = q.deconstruct()
-        assert args == (
-            ('price__gt', 10.0),
-            ('price', 11.0),
-        )
-        assert kwargs == {}
-
-    def test_deconstruct_multiple_kwargs(self):
-        q = Q(price__gt=10.0, price=11.0)
-        path, args, kwargs = q.deconstruct()
-        assert args == (
-            ('price', 11.0),
-            ('price__gt', 10.0),
-        )
-        assert kwargs == {}
-
-    def test_reconstruct(self):
-        q = Q(price__gt=10.0)
-        path, args, kwargs = q.deconstruct()
-        assert Q(*args, **kwargs) == q
-
-    def test_reconstruct_negated(self):
-        q = ~Q(price__gt=10.0)
-        path, args, kwargs = q.deconstruct()
-        assert Q(*args, **kwargs) == q
-
-    def test_reconstruct_or(self):
-        q1 = Q(price__gt=10.0)
-        q2 = Q(price=11.0)
-        q = q1 | q2
-        path, args, kwargs = q.deconstruct()
-        assert Q(*args, **kwargs) == q
-
-    def test_reconstruct_and(self):
-        q1 = Q(price__gt=10.0)
-        q2 = Q(price=11.0)
-        q = q1 & q2
-        path, args, kwargs = q.deconstruct()
-        assert Q(*args, **kwargs) == q
-
-
-class TestConjunctions:
-    """Class that holds tests cases for Conjunctions (AND, OR, NeG)"""
-
-    def test_default_AND(self):
-        """Test that kwargs to `filter` are ANDed by default"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        q1 = Dog.query.filter(owner='John', age=3)
-        assert q1.total == 1
-
-    def test_default_NEG_AND(self):
-        """Test that kwargs to `filter` are ANDed by default"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        q1 = Dog.query.exclude(owner='John', age=3)
-        assert q1.total == 1
-
-        q2 = Dog.query.exclude(owner='Carrie', age=10)
-        assert q2.total == 2
-
-    def test_simple_AND(self):
-        """Test straightforward AND of two criteria"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        # Filter by the Owner
-        q1 = Dog.query.filter(Q(owner='John') & Q(age=3))
-        assert q1.total == 1
-
-    def test_simple_OR(self):
-        """Test straightforward OR of two criteria"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        q1 = Dog.query.filter(Q(owner='John') | Q(age=3))
-        assert q1.total == 2
-
-    def test_AND_with_OR(self):
-        """Test combination of AND and OR"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-        Dog.create(id=5, name='Leslie', age=6, owner='Underwood')
-        Dog.create(id=6, name='Dave', age=6, owner='Carrie')
-
-        q1 = Dog.query.filter(
-            Q(owner='John', name='Jean') |
-            Q(age=6))
-        assert q1.total == 4
-
-        q2 = Dog.query.filter(Q(owner='John') | Q(age=6))
-        assert q2.total == 5
-
-        q3 = Dog.query.filter(
-            (Q(owner='John') & Q(age=7)) |
-            (Q(owner='Carrie') & Q(age=6)))
-        assert q3.total == 3
-
-    def test_OR_with_AND(self):
-        """Test combination of OR and AND"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-        Dog.create(id=5, name='Leslie', age=6, owner='Underwood')
-        Dog.create(id=6, name='Dave', age=6, owner='Carrie')
-
-        q1 = Dog.query.filter((Q(owner='John') | Q(age=7)) & (Q(owner='Carrie') | Q(age=6)))
-        assert q1.total == 0
-
-        q2 = Dog.query.filter(
-            (Q(owner='John') | Q(age__gte=3)) &
-            (Q(name='Jean') | Q(name='Murdock')))
-        assert q2.total == 2
-
-    def test_NEG(self):
-        """Test Negation of a criteria"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-        Dog.create(id=5, name='Leslie', age=6, owner='Underwood')
-        Dog.create(id=6, name='Dave', age=6, owner='Carrie')
-
-        q1 = Dog.query.filter(~Q(owner='John'))
-        assert q1.total == 3
-
-        q2 = Dog.query.filter(~Q(owner='John') | ~Q(age=7))
-        assert q2.total == 4
-
-    def test_empty_resultset(self):
-        """Test that kwargs to `filter` are ANDed by default"""
-        # Add multiple entries to the DB
-        Dog.create(id=2, name='Murdock', age=7, owner='John')
-        Dog.create(id=3, name='Jean', age=3, owner='John')
-        Dog.create(id=4, name='Bart', age=6, owner='Carrie')
-
-        q1 = Dog.query.filter(owner='XYZ', age=100)
-        assert q1.total == 0
-
-
-class TestAssociations:
-    """Class that holds tests cases for Entity Associations"""
-
-    class TestReference:
-        """Class to test References (Foreign Key) Association"""
-
-        def test_init(self):
-            """Test successful RelatedDog initialization"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-            assert dog.owner.id == human.id
-            assert dog.owner_id == human.id
-
-        def test_init_with_string_reference(self):
-            """Test successful RelatedDog initialization"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog2(id=1, name='John Doe', age=10, owner=human)
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-            assert dog.owner.id == human.id
-            assert dog.owner_id == human.id
-            assert not hasattr(human, 'dog')  # Reverse linkages are not provided by default
-
-        def test_save(self):
-            """Test successful RelatedDog save"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-            dog.save()
-            assert dog.id is not None
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-
-        def test_unsaved_entity_init(self):
-            """Test that initialization fails when an unsaved entity is assigned to a relation"""
-            with pytest.raises(ValueError):
-                human = Human(first_name='Jeff', last_name='Kennedy', email='jeff.kennedy@presidents.com')
-                RelatedDog(id=1, name='John Doe', age=10, owner=human)
-
-        def test_unsaved_entity_assign(self):
-            """Test that assignment fails when an unsaved entity is assigned to a relation"""
-            with pytest.raises(ValueError):
-                human = Human(first_name='Jeff', last_name='Kennedy',
-                              email='jeff.kennedy@presidents.com')
-
-                dog = RelatedDog(id=1, name='John Doe', age=10)
-                assert any(key in dog.__dict__ for key in ['owner', 'owner_id']) is False
-                dog.owner = human
-
-        def test_invalid_entity_type(self):
-            """Test that assignment fails when an invalid entity type is assigned to a relation"""
-            with pytest.raises(ValidationError):
-                dog = Dog.create(name='Johnny', owner='John')
-                related_dog = RelatedDog(id=1, name='John Doe', age=10)
-                related_dog.owner = dog
-
-        def test_shadow_attribute(self):
-            """Test identifier backing the association"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-            assert human.id is not None
-            assert dog.owner_id == human.id
-
-        def test_save_after_assign(self):
-            """Test saving after assignment (post init)"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10)
-            assert any(key in dog.__dict__ for key in ['owner', 'owner_id']) is False
-            dog.owner = human
-            dog.save()
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-            assert dog.owner_id == human.id
-
-        def test_fetch_after_save(self):
-            """Test fetch after save and ensure reference is not auto-loaded"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10)
-            dog.owner = human
-            dog.save()
-
-            dog2 = RelatedDog.get(dog.id)
-            # Reference attribute is not loaded automatically
-            assert 'owner' not in dog2.__dict__
-            assert dog2.owner_id == human.id
-
-            # Accessing attribute shows it up in __dict__
-            getattr(dog2, 'owner')
-            assert 'owner' in dog2.__dict__
-
-        def test_shadow_attribute_init(self):
-            """Test identifier backing the association"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner_id=human.id)
-            assert 'owner_id' in dog.__dict__
-            assert 'owner' not in dog.__dict__
-            dog.save()
-            assert dog.owner_id == human.id
-            assert dog.owner.id == human.id
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_id'])
-
-        def test_shadow_attribute_assign(self):
-            """Test identifier backing the association"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10)
-            dog.owner_id = human.id
-            assert 'owner' not in dog.__dict__
-            dog.save()
-            assert dog.owner_id == human.id
-            assert dog.owner.id == human.id
-            assert 'owner' in dog.__dict__
-
-        def test_reference_reset_association_to_None(self):
-            """Test that the reference field and shadow attribute are reset together"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert dog.owner_id == human.id
-            assert dog.owner.id == human.id
-
-            dog.owner = None
-            assert any(key in dog.__dict__ for key in ['owner', 'owner_id']) is False
-            assert dog.owner is None
-            assert dog.owner_id is None
-
-        def test_reference_reset_shadow_field_to_None(self):
-            """Test that the reference field and shadow attribute are reset together"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert dog.owner_id == human.id
-            assert dog.owner.id == human.id
-
-            dog.owner_id = None
-            assert any(key in dog.__dict__ for key in ['owner', 'owner_id']) is False
-            assert dog.owner is None
-            assert dog.owner_id is None
-
-        def test_reference_reset_association_by_del(self):
-            """Test that the reference field and shadow attribute are reset together"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert dog.owner_id == human.id
-            assert dog.owner.id == human.id
-
-            del dog.owner
-            assert any(key in dog.__dict__ for key in ['owner', 'owner_id']) is False
-            assert dog.owner is None
-            assert dog.owner_id is None
-
-        def test_reference_reset_shadow_field_by_del(self):
-            """Test that the reference field and shadow attribute are reset together"""
-            human = Human.create(id=101, first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner=human)
-            assert dog.owner_id == human.id
-            assert dog.owner.id == human.id
-
-            del dog.owner_id
-            assert any(key in dog.__dict__ for key in ['owner', 'owner_id']) is False
-            assert dog.owner is None
-            assert dog.owner_id is None
-
-        def test_via(self):
-            """Test successful save with an entity linked by via"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = DogRelatedByEmail.create(id=1, name='John Doe', age=10, owner=human)
-            assert all(key in dog.__dict__ for key in ['owner', 'owner_email'])
-            assert hasattr(dog, 'owner_email')
-            assert dog.owner_email == human.email
-
-        def test_via_with_shadow_attribute_assign(self):
-            """Test successful save with an entity linked by via"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = DogRelatedByEmail(id=1, name='John Doe', age=10)
-            dog.owner_email = human.email
-            assert 'owner' not in dog.__dict__
-            dog.save()
-            assert hasattr(dog, 'owner_email')
-            assert dog.owner_email == human.email
-
-        @mock.patch('protean.core.entity.Entity.find_by')
-        def test_caching(self, find_by_mock):
-            """Test that subsequent accesses after first retrieval don't fetch record again"""
-            human = Human.create(first_name='Jeff', last_name='Kennedy',
-                                 email='jeff.kennedy@presidents.com')
-            dog = RelatedDog(id=1, name='John Doe', age=10, owner_id=human.id)
-
-            for _ in range(3):
-                getattr(dog, 'owner')
-            assert find_by_mock.call_count == 1
-
-    class TestHasOne:
-        """Class to test HasOne Association"""
-
-        def test_init(self):
-            """Test successful HasOne initialization"""
-            human = HasOneHuman1.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            dog = HasOneDog1.create(id=101, name='John Doe', age=10, has_one_human1=human)
-            assert dog.has_one_human1 == human
-            assert human.dog.id == dog.id
-
-        def test_init_with_via(self):
-            """Test successful HasOne initialization with a class containing via"""
-            human = HasOneHuman2.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            dog = HasOneDog2.create(id=101, name='John Doe', age=10, human=human)
-            assert dog.human == human
-            assert 'dog' not in human.__dict__
-            assert human.dog.id == dog.id
-            assert 'dog' in human.__dict__
-
-        def test_init_with_no_reference(self):
-            """Test successful HasOne initialization with a class containing via"""
-            human = HasOneHuman3.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            dog = HasOneDog3.create(id=101, name='John Doe', age=10, human_id=human.id)
-            assert dog.human_id == human.id
-            assert not hasattr(dog, 'human')
-            assert 'human' not in dog.__dict__
-            assert human.dog.id == dog.id
-
-        @mock.patch('protean.core.entity.Entity.find_by')
-        def test_caching(self, find_by_mock):
-            """Test that subsequent accesses after first retrieval don't fetch record again"""
-            human = HasOneHuman3.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            HasOneDog3.create(id=101, name='John Doe', age=10, human_id=human.id)
-
-            for _ in range(3):
-                getattr(human, 'dog')
-            assert find_by_mock.call_count == 1
-
-    class TestHasMany:
-        """Class to test HasMany Association"""
-
-        def test_init(self):
-            """Test successful HasOne initialization"""
-            human = HasManyHuman1.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            dog1 = HasManyDog1.create(id=101, name='John Doe', age=10, has_many_human1=human)
-            dog2 = HasManyDog1.create(id=102, name='Jane Doe', age=10, has_many_human1=human)
-
-            assert dog1.has_many_human1.id == human.id
-            assert dog2.has_many_human1.id == human.id
-            assert 'dogs' not in human.__dict__
-            assert len(human.dogs) == 2
-            assert 'dogs' in human.__dict__  # Avaiable after access
-
-        def test_init_with_via(self):
-            """Test successful HasMany initialization with a class containing via"""
-            human = HasManyHuman2.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            dog1 = HasManyDog2.create(id=101, name='John Doe', age=10, human=human)
-            dog2 = HasManyDog2.create(id=102, name='Jane Doe', age=10, human=human)
-
-            assert dog1.human.id == human.id
-            assert dog2.human.id == human.id
-
-            assert len(human.dogs) == 2
-
-        def test_init_with_no_reference(self):
-            """Test successful HasOne initialization with a class containing via"""
-            human = HasManyHuman3.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            dog1 = HasManyDog3.create(id=101, name='John Doe', age=10, human_id=human.id)
-
-            assert dog1.human_id == human.id
-            assert not hasattr(dog1, 'human')
-
-        @mock.patch('protean.core.entity.QuerySet.filter')
-        @mock.patch('protean.core.entity.Entity.exists')
-        def test_caching(self, exists_mock, filter_mock):
-            """Test that subsequent accesses after first retrieval don't fetch record again"""
-            exists_mock.return_value = False
-            human = HasManyHuman3.create(
-                first_name='Jeff', last_name='Kennedy',
-                email='jeff.kennedy@presidents.com')
-            HasManyDog3.create(id=101, name='John Doe', human_id=human.id)
-            HasManyDog3.create(id=102, name='Jane Doe', human_id=human.id)
-            HasManyDog3.create(id=103, name='Jinny Doe', human_id=human.id)
-
-            for _ in range(3):
-                getattr(human, 'dogs')
-            assert filter_mock.call_count == 1
